add RemoteConnectionsBehind which inherits ObservableBehind
revamp ConnectionStatus component to Observe the IConnectionsCounter
implemented by the new behind

https://github.com/Not-Endurance/not-timing-system/issues/336
