+++ src/Judge/NTS.Judge.MAUI/MauiProgram.cs
@@ -34,7 +34,7 @@ public static class MauiProgram
             var currentDirectory = Directory.GetCurrentDirectory();
             var info = new ProcessStartInfo
             {
-                //FileName = Path.Combine(currentDirectory, "NTS.Judge.MAUI.Server.exe"),
+                FileName = Path.Combine(currentDirectory, "NTS.Judge.MAUI.Server.exe"),
             };
 
             var hubProcess = Process.Start(info);
