--- src/Judge/NTS.Judge/Core/Behinds/RemoteConnectionsCount.cs
@@ -1,23 +1,0 @@
-﻿using Not.Application.RPC;
-
-namespace NTS.Judge.Core.Behinds;
-
-public class RemoteConnectionsCount : IConnectionsCounter
-{
-    public List<string> ActiveConnections { get; } = [];
-
-    public void AddConnection(string connectionId)
-    {
-        ActiveConnections.Add(connectionId);
-    }
-
-    public void RemoveConnection(string connectionId)
-    {
-        ActiveConnections.Remove(connectionId);
-    }
-
-    public int GetConnectionsCount()
-    {
-        return ActiveConnections.Count;
-    }
-}
